added avatar

+++ src/api/api.js
@@ -1,36 +1,66 @@
 import axios from 'axios';
 import jwt_decode from 'jwt-decode';
-export const backendUrl = 'http://192.168.0.110:5000';
 
+export const backendUrl = 'http://localhost:5000';
 class ApiService {
   httpGET(url) {
     if (this.validateToken()) {
       return axios.get(backendUrl + url, this.authorize())
         .then(this.mapResponse);
     } else {
-      window.location.reload(false);
-      
+      return this.ifExpired()
+    }
+  }
+
+  httpGETAvatar(url) {
+    if (this.validateToken()) {
+      return axios.get(backendUrl + url, this.authorizeWithBlob())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
     }
   }
 
   httpPOST(url, data = {}) {
+    if (this.validateToken()) {
+      return axios.post(backendUrl + url, data, this.authorize())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
+    }
+  }
+
+  httpAuthPOST(url, data = {}) {
     return axios.post(backendUrl + url, data, this.authorize())
       .then(this.mapResponse);
   }
 
   httpDELETE(url) {
-    return axios.delete(backendUrl + url, this.authorize())
-      .then(this.mapResponse);
+    if (this.validateToken()) {
+      return axios.delete(backendUrl + url, this.authorize())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
+    }
   }
 
   httpPUT(url, data = {}) {
-    return axios.put(backendUrl + url, data, this.authorize())
-      .then(this.mapResponse);
+    if (this.validateToken()) {
+      return axios.put(backendUrl + url, data, this.authorize())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
+    }
   }
 
+
   httpPATCH(url, data = {}) {
-    return axios.patch(backendUrl + url, data, this.authorize())
-      .then(this.mapResponse);
+    if (this.validateToken()) {
+      return axios.patch(backendUrl + url, data, this.authorize())
+        .then(this.mapResponse);
+    } else {
+      return this.ifExpired()
+    }
   }
 
   mapResponse = (response) => response.data
@@ -38,7 +68,24 @@ class ApiService {
   authorize = () => {
     const token = JSON.parse(localStorage.getItem('token'));
     if (token) {
-      return { headers: { Authorization: 'Bearer ' + token } };
+      return {
+        headers: {
+          Authorization: 'Bearer ' + token,
+        }
+      };
+    }
+    return null;
+  }
+
+  authorizeWithBlob = () => {
+    const token = JSON.parse(localStorage.getItem('token'));
+    if (token) {
+      return {
+        responseType: 'blob',
+        headers: {
+          Authorization: 'Bearer ' + token,
+        }
+      };
     }
     return null;
   }
@@ -49,11 +96,19 @@ class ApiService {
     if (decodedToken.exp * 1000 < currentDate.getTime()) {
       console.log("Token expired.");
       localStorage.removeItem('token')
+      localStorage.setItem('tokenExpired', true);
       return false;
     } else {
       console.log("Valid token");
       return true;
     }
+  }
+
+  ifExpired = () => {
+    window.location.reload(false);
+    return new Promise((resolve) => {
+      resolve(console.log("Token expired, redirect to default page"));
+    })
   }
 }
